Add validation for @Provides methods found in objects
Currently this is not supported and @Provides must live within a class, interface, or companion object. Previously, when migrating from Dagger+Anvil, non-contributed @Modules defined as objects result in a NullPointerException in IR for the ProvidesTransformer. These changes make it so we catch the situation earlier and present an explicit error message pointing to the class+method in question.

Also update the error message for top-level @Provides/@Binds functions to be more explicit.