Fix @ElementsIntoSet multibinding contributions triggering a dependency cycle in some situations
The dependency cycle would primarily occur when contributing elements into set in a contributed interface, and that multibinding gets injected as a constructor argument. Changes that would work-around the issue included using @IntoSet instead of providing the set of elements, or accessing the multibinding directly as a dependency graph-level provider field.