attempt to exclude FirPropertyAccessor with a failing scenario in test
It is interesting that both `special/accessor` and `<local>/field` have `Source` origin, it means we cannot exclude them by checking Keys.

I tried to exclude FirPropertyAccessor as a whole, but looks like it breaks a scenario (see the diff I added in this commit), it looks like a valid use case (?) and I don't want to break it here.